Move some greet tests to gmsh tests.
